migrate fundrazr model logic
