+++ microsetta_private_api/model/tests/test_campaign.py
